funcionando planes de compensacion

+++ carbon_foot_print/src/main/java/com/cruzroja/carbon_foot_print/Infrastucture/Input/ControllerManageCompensationPlan/DTOResponse/CompensationPlanDTOResponse.java
@@ -15,7 +15,7 @@ public class CompensationPlanDTOResponse {
     private String planName;
     private double planPrice;
     private String planDescription;
-    private int planDiscount;
+    private double planDiscount;
     private List<ActionDTOResponse> actions;
 
     public CompensationPlanDTOResponse(){
