+++ carbon_foot_print/src/main/java/com/cruzroja/carbon_foot_print/Infrastucture/Output/Persistence/Entities/CompensationPlanEntity.java
@@ -35,7 +35,7 @@ public class CompensationPlanEntity {
     private String planDescription;
 
     @Column(nullable = false)
-    private int planDiscount;
+    private double planDiscount;
 
     @ManyToMany(fetch = FetchType.EAGER)
     @JoinTable(name = "plans_actions",
